requirements.txt
pandas
numpy
scikit-learn
pathlib
mlflow
python-dotenv